scripts/requirements.txt
Pillow
transformers
timm
pytesseract
opencv-python
nltk
keras
tensorflow
scipy
scikit-learn
psycopg
asyncpg
fastapi
"uvicorn[standard]"
python-multipart
aiofiles
requests
click
PyMuPDF-1.21.1
